+++ multiqc/core/write_results.py
@@ -10,9 +10,9 @@ import subprocess
 import sys
 import time
 import traceback
+import uuid
 from pathlib import Path
 from typing import Optional
-import uuid
 
 import jinja2