+++ multiqc/plots/plot.py
@@ -28,7 +28,7 @@ from multiqc import config, report
 from multiqc.core import tmp_dir
 from multiqc.core.strict_helpers import lint_error
 from multiqc.plots.utils import check_plotly_version
-from multiqc.types import Anchor, PlotType
+from multiqc.types import Anchor, PlotType, SampleName
 from multiqc.utils import mqc_colour
 from multiqc.validation import ValidatedConfig, add_validation_warning
 
@@ -215,6 +215,9 @@ class BaseDataset(BaseModel):
     trace_params: Dict[str, Any]
     pct_range: Dict[str, Any]
     n_samples: int
+
+    def samples_names(self) -> List[SampleName]:
+        raise NotImplementedError
 
     def create_figure(
         self,
@@ -1056,7 +1059,6 @@ class Plot(BaseModel, Generic[DatasetT, PConfigT]):
                     type="button"
                     data-toggle="tooltip" 
                     aria-controls="{section_anchor}_ai_summary_wrapper"
-                    title="Dynamically generate AI summary for this plot"
                 >
                     <span style="vertical-align: baseline">
                         <svg width="11" height="10" viewBox="0 0 17 15" fill="black" xmlns="http://www.w3.org/2000/svg">