+++ multiqc/plots/violin.py
@@ -528,6 +528,7 @@ class Dataset(BaseDataset):
                 for _, _, col in headers
             ):
                 continue
+            pseudonym = report.anonymize_sample_name(sample)
             row = []
             for _, _, col in headers:
                 value = self.violin_value_by_sample_by_metric[col.rid].get(
@@ -537,7 +538,7 @@ class Dataset(BaseDataset):
                     if col.suffix:
                         value = f"{value}{col.suffix}"
                 row.append(str(value))
-            result += f"| {sample} | " + " | ".join(row) + " |\n"
+            result += f"| {pseudonym} | " + " | ".join(row) + " |\n"
 
         return result
 
@@ -551,6 +552,12 @@ class ViolinPlot(Plot[Dataset, TableConfig]):
     show_table_by_default: bool
     n_samples: int
     table_anchor: Anchor
+
+    def samples_names(self) -> List[SampleName]:
+        names: List[SampleName] = []
+        for ds in self.datasets:
+            names.extend(SampleName(s) for s in ds.all_samples)
+        return names
 
     # @staticmethod
     # def create(